update the interface of init_binning_fourier, set lmin/lmax to float

+++ Cocoa/external_modules/code/cosmolike/generic_interface.cpp
@@ -466,8 +466,8 @@ void init_binning_cmb_bandpower(
 
 void init_binning_fourier(
     const int Nells, 
-    const int lmin, 
-    const int lmax
+    const double lmin, 
+    const double lmax
   )
 {
   spdlog::debug("{}: Begins", "init_binning_fourier");
